old-tests/submission/W3C/domhtml/HTMLDocument03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/origin-0.html#dom-document-domain" rel="help">
<link rel="author" title="W3C">
</head>
<body id="TEST-BODY"> 
<p>The domain attribute specifies the domain name of the server that served
    the document, or null if the server cannot be identified by a domain name.

    Retrieve the domain attribute and examine its value.</p>

<script>
var domtest = async_test("The domain attribute specifies the domain name of the server that served     the document, or null if the server cannot be identified by a domain name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vdomain;
      var doc;
      doc = window.document;
      vdomain = doc.domain;

      var url = doc.URL;
      if (url.indexOf("file:") != -1) {
        assert_equals(vdomain,null);
      } else {
        var start = url.indexOf("//");
        var end = url.substring(start+2).indexOf('/');
        var server = url.substring(start+2, end+start+2);
        var colon = server.indexOf(':');
        if (colon != -1) server = server.substring(0, colon);
        assert_equals(vdomain,server);
      }
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
